# docs/swagger.yaml
basePath: /
definitions:
  simplesql.AddProductParams:
    properties:
      name:
        type: string
      price:
        type: integer
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: twp server api.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: twp API
  version: 0.o
paths:
  /file:
    post:
      consumes:
      - multipart/form-data
      description: Upload a file to the server
      parameters:
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: json
      summary: Upload a file
      tags:
      - Files
  /file/{file_name}:
    get:
      description: Redirect the user to the URL of the specified file
      parameters:
      - description: Name of the file to retrieve
        in: path
        name: file_name
        required: true
        type: string
      responses:
        "301":
          description: Moved Permanently
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Get a file
      tags:
      - Files
  /product:
    post:
      consumes:
      - application/json
      description: Add a new product to the database
      parameters:
      - description: Product details
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/simplesql.AddProductParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Add a new product
      tags:
      - Products
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"